@@ -20,16 +20,16 @@
   ],
   "license": "MIT",
   "devDependencies": {
-    "@types/jest": "^25.1.4",
-    "@types/node": "^16.18.0",
-    "jest": "^25.1.0",
-    "ts-jest": "^25.2.1",
-    "typescript": "^3.9.7"
+    "@types/jest": "^29.5.12",
+    "@types/node": "^20.11.30",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.1.4",
+    "typescript": "^5.4.3"
   },
   "dependencies": {
-    "@actions/core": "^1.10.0",
-    "@types/q": "^1.5.2",
-    "azure-actions-appservice-rest": "^1.3.11",
+    "@actions/core": "^1.10.1",
+    "@types/q": "^1.5.8",
+    "azure-actions-appservice-rest": "^1.3.13",
     "azure-actions-webclient": "^1.1.1"
   }
 }
