@@ -19,8 +19,8 @@ class ContainerDeploymentUtility {
         this._appService = appService;
         this._appServiceUtility = new AzureAppServiceUtility_1.AzureAppServiceUtility(appService);
     }
-    deployWebAppImage(images, multiContainerConfigFile, isLinux, isMultiContainer, startupCommand, restart = true) {
-        return __awaiter(this, void 0, void 0, function* () {
+    deployWebAppImage(images_1, multiContainerConfigFile_1, isLinux_1, isMultiContainer_1, startupCommand_1) {
+        return __awaiter(this, arguments, void 0, function* (images, multiContainerConfigFile, isLinux, isMultiContainer, startupCommand, restart = true) {
             let updatedMulticontainerConfigFile = multiContainerConfigFile;
             if (isMultiContainer) {
                 core.debug("Deploying Docker-Compose file " + multiContainerConfigFile + " to the webapp " + this._appService.getName());
@@ -85,7 +85,7 @@ class ContainerDeploymentUtility {
             if (isLinuxApp) {
                 if (isMultiContainer) {
                     let fileData = fs.readFileSync(multicontainerConfigFile);
-                    appSettingsNewProperties["linuxFxVersion"] = "COMPOSE|" + (new Buffer(fileData).toString('base64'));
+                    appSettingsNewProperties["linuxFxVersion"] = "COMPOSE|" + (Buffer.from(fileData).toString('base64'));
                 }
                 else {
                     appSettingsNewProperties["linuxFxVersion"] = "DOCKER|" + imageName;
